Fix types of timeout and usb_transfer_size

+++ include/larpix.h
@@ -10,8 +10,8 @@ typedef struct larpix_connection
     uint clk_divisor;
     byte pin_io_directions;
     byte bit_mode;
-    byte timeout;
-    byte usb_transfer_size;
+    uint timeout;
+    uint usb_transfer_size;
 } larpix_connection;
 
 void larpix_default_connection(larpix_connection* c);
